Store data restructured and working labels adding system
Sooo i rewrite reducers to make them more readable. It's still not best so in future i will change it futher on.

Added working way to add labels to note and display them

Now i'm going to work on solution to rename labels because i need to rename either labels in store but also in notes